Fix: UserLikedPosts return successful
Bug fix: UserLikedPosts were not returning the posts.

Some refactoring of tests too.

Note: Some code is commented out and will be used in new functions after rearrangement. Do not delete it.
