+++ backend/server/handlers_posts_category.go
@@ -6,26 +6,21 @@ import (
 )
 
 // postsCategoriesHandler returns a json list of all categories from the database
+//
+// Example: /api/posts/categories
 func (srv *Server) postsCategoriesHandler(w http.ResponseWriter, _ *http.Request) {
 	sendObject(w, categories)
 }
 
+// postsCategoriesNameHandler returns a json list of all posts belonging to the specified category
+//
+// Example: /api/posts/categories/name
 func (srv *Server) postsCategoriesNameHandler(w http.ResponseWriter, r *http.Request) {
 	categoryName := strings.TrimPrefix(r.URL.Path, "/api/posts/categories/")
-	// /api/posts/categories/name -> name
 
 	categoryName = strings.ToLower(categoryName)
-	// Name -> name
 
-	isValid := false
-	for _, cat := range categories {
-		if cat == categoryName {
-			isValid = true
-			break
-		}
-	}
-
-	if !isValid {
+	if !isPresent(categories, categoryName) {
 		errorResponse(w, http.StatusNotFound)
 		return
 	}
